+++ src/ubigint/sub.rs
@@ -19,7 +19,7 @@ impl<const N_BITS: usize> UBigIntImpl<N_BITS> {
 
             // from     A1      - (B1        + borrow_0)
             //   to     A{N-2}  - (B{N-2}    + borrow_{N-3})
-            for _ in 0..(n_limbs - 2) as u32 {
+            for _ in 0..n_limbs - 2 {
                 OP_ROT
                 OP_ADD
                 OP_SWAP
@@ -33,7 +33,7 @@ impl<const N_BITS: usize> UBigIntImpl<N_BITS> {
             OP_ADD
             { u30_sub_nocarry(head_offset) }
 
-            for _ in 0..(n_limbs - 1) as u32 {
+            for _ in 0..n_limbs - 1 {
                 OP_FROMALTSTACK
             }
         }
